pyproject.toml
[tool.poetry]
name = "hb-bot"
version = "0.1.0"
description = ""
authors = ["Vadim <explosive.va@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
